Refactor: move BMI calculations/data
